requirements.txt
attrs==22.2.0
exceptiongroup==1.1.0
iniconfig==2.0.0
mysql-connector-python==8.0.32
mysqlclient==2.1.1
packaging==23.0
peewee==3.15.4
pluggy==1.0.0
protobuf==3.20.3
pytest==7.2.1
tomli==2.0.1
